[OpenXR] Use a WorkQueue for the render loop https://bugs.webkit.org/show_bug.cgi?id=297451
Reviewed by NOBODY (OOPS!).

We currently use a thread to poll the events and process the frame, but
several objects are created in main thread and used from both threads or
can be released in both threads too. Using a WorkQueue we can move all
the session handling to the render thread and simplify the render loop.