[WPE] Crash if unable to create a renderer backend https://bugs.webkit.org/show_bug.cgi?id=298775
Reviewed by NOBODY (OOPS!).

PlatformDisplayLibWPE::create aborts if it is unable to create a
glDisplay, except in the case where it is unable to create a renderer
backend, in which case it silently returns a nullptr.

This patch adds a crash in that scenario and prints the file descriptor
for which that happens.

* Source/WebCore/platform/graphics/libwpe/PlatformDisplayLibWPE.cpp:
(WebCore::PlatformDisplayLibWPE::create):